refactor: movendo componentes para src/
