+++ mobile/app/(rep)/index.tsx
@@ -1,6 +1,6 @@
 import { View, Text, StyleSheet, ScrollView, SafeAreaView } from 'react-native';
-import { VisitCard } from '../../components/VisitCard'; 
-import { Header } from '../../components/Header'; 
+import { VisitCard } from '../../src/components/VisitCard'; 
+import { Header } from '../../src/components/Header'; 
 import { colors } from '../../src/theme/colors';
 import { useRouter } from 'expo-router';
